+++ python/mocpy/moc/moc.py
@@ -986,7 +986,7 @@ class MOC(AbstractMOC):
 
     @classmethod
     @validate_lonlat
-    def from_elliptical_cone(cls, lon, lat, a, b, pa, max_depth, delta_depth=2):
+    def from_elliptical_cone(cls, lon, lat, a, b, pa, max_depth, *, delta_depth=2):
         """
         Create a MOC from an elliptical cone.
 
@@ -1046,7 +1046,7 @@ class MOC(AbstractMOC):
 
     @classmethod
     @validate_lonlat
-    def from_cone(cls, lon, lat, radius, max_depth, delta_depth=2):
+    def from_cone(cls, lon, lat, radius, max_depth, *, delta_depth=2):
         """
         Create a MOC from a cone.
 
@@ -1091,9 +1091,73 @@ class MOC(AbstractMOC):
             lat[0],
             np.float64(radius.to_value(u.deg)),
             np.uint8(max_depth),
-            np.uint8(delta_depth),
+            delta_depth=np.uint8(delta_depth),
         )
         return cls(index)
+
+    @classmethod
+    @validate_lonlat
+    def from_cones(cls, lon, lat, radius, max_depth, *, delta_depth=2, n_threads=None):
+        """
+        Create a list of MOCs from cones.
+
+        Each cone is centered around the (`lon`, `lat`) position with a radius expressed by
+        `radius`. Coordinates should be in the ICRS frame, as stipulated in the MOC standard.
+
+        Parameters
+        ----------
+        lon : `astropy.coordinates.Longitude` or its supertype `astropy.units.Quantity`
+            The longitude of the center of the cone. Can be scalar or a list of longitudes.
+        lat : `astropy.coordinates.Latitude` or its supertype `astropy.units.Quantity`
+            The latitude of the center of the cone. Can be scalar or a list of latitudes.
+        radius : `astropy.coordinates.Angle` or its supertype `astropy.units.Quantity`
+            The radius angle of the cone. Can be scalar or a list of radii.
+        max_depth : int
+            Maximum HEALPix cell resolution.
+        delta_depth : int, optional
+            To control the approximation, you can choose to perform the computations at a deeper
+            depth using the `depth_delta` parameter.
+            The depth at which the computations will be made will therefore be equal to
+            `max_depth` + `depth_delta`.
+        n_threads : int, optional
+            The number of threads to be used. If this is set to None (default value),
+            all available threads will be used.
+
+        Returns
+        -------
+        result : List[`~mocpy.moc.MOC`]
+            The resulting list of MOCs
+
+        Examples
+        --------
+        >>> from mocpy import MOC
+        >>> import astropy.units as u
+        >>> moc = MOC.from_cone(
+        ...  lon=[1, 4] * u.deg,
+        ...  lat=[2, 5] * u.deg,
+        ...  radius=10 * u.deg,
+        ...  max_depth=10
+        ... )
+        """
+        if radius.isscalar:
+            indices = mocpy.from_same_cones(
+                lon,
+                lat,
+                np.float64(Angle(radius).to_value(u.deg)),
+                np.uint8(max_depth),
+                delta_depth=np.uint8(delta_depth),
+                n_threads=n_threads,
+            )
+            return [cls(index) for index in indices]
+        indices = mocpy.from_cones(
+            lon,
+            lat,
+            np.float64(Angle(radius).to_value(u.deg)),
+            np.uint8(max_depth),
+            delta_depth=np.uint8(delta_depth),
+            n_threads=n_threads,
+        )
+        return [cls(index) for index in indices]
 
     @classmethod
     @validate_lonlat
@@ -1145,6 +1209,90 @@ class MOC(AbstractMOC):
             np.uint8(max_depth),
         )
         return cls(index)
+
+    @classmethod
+    @validate_lonlat
+    def from_boxes(cls, lon, lat, a, b, angle, max_depth, *, n_threads=None):
+        """
+        Create a MOC from a box/rectangle.
+
+        The box is centered around the (`lon`, `lat`) position.
+
+        Parameters
+        ----------
+        lon : `~astropy.coordinates.Longitude` or its supertype `~astropy.units.Quantity`
+            The longitude of the center of the cone.
+        lat : `~astropy.coordinates.Latitude` or its supertype `~astropy.units.Quantity`
+            The latitude of the center of the cone.
+        a : `~astropy.coordinates.Angle` or its supertype `~astropy.units.Quantity`
+            The semi-major axis of the box, i.e. half of the box's width.
+        b : `~astropy.coordinates.Angle` or its supertype `~astropy.units.Quantity`
+            The semi-minor axis of the box, i.e. half of the box's height.
+        angle : `astropy.coordinates.Angle` or its supertype `~astropy.units.Quantity`
+            The tilt angle between the north and the semi-major axis, east of north.
+        max_depth : int
+            Maximum HEALPix cell resolution.
+        n_threads : int, optional
+            The number of threads to be used. If this is set to None (default value),
+            all available threads will be used.
+
+        Returns
+        -------
+        result : list[`~mocpy.moc.MOC`]
+            The resulting list of MOCs.
+
+        Examples
+        --------
+        >>> from mocpy import MOC
+        >>> import astropy.units as u
+        >>> # similar boxes, same size and orientation
+        >>> moc_list = MOC.from_boxes(
+        ...  lon=[1, 2]*u.deg,
+        ...  lat=[1, 2]*u.deg,
+        ...  a=10*u.deg,
+        ...  b=5*u.deg,
+        ...  angle=30*u.deg,
+        ...  max_depth=10
+        ... )
+        >>> # different boxes
+        >>> moc_list = MOC.from_boxes(
+        ...  lon=[1, 2]*u.deg,
+        ...  lat=[1, 2]*u.deg,
+        ...  a=[10, 20]*u.deg,
+        ...  b=[5, 10]*u.deg,
+        ...  angle=[30, 10]*u.deg,
+        ...  max_depth=10
+        ... )
+        """
+        params = [a, b, angle]
+        if any(isinstance(param, u.Quantity) and param.isscalar for param in params):
+            if not all(isinstance(param, u.Quantity) for param in params):
+                raise ValueError(
+                    "'a', 'b' and 'angle' should either be all astropy angle-equivalent"
+                    "scalar values or they should all be iterable angle-equivalent. "
+                    "They cannot be a mix of both.",
+                )
+            indices = mocpy.from_same_boxes(
+                lon,
+                lat,
+                np.float64(a.to_value(u.deg)),
+                np.float64(b.to_value(u.deg)),
+                np.float64(angle.to_value(u.deg)),
+                np.uint8(max_depth),
+                n_threads=n_threads,
+            )
+            return [cls(index) for index in indices]
+        # different boxes
+        indices = mocpy.from_boxes(
+            lon,
+            lat,
+            np.array(Angle(a).to_value(u.deg), dtype=np.float64),
+            np.array(Angle(b).to_value(u.deg), dtype=np.float64),
+            np.array(Angle(angle).to_value(u.deg), dtype=np.float64),
+            np.uint8(max_depth),
+            n_threads=n_threads,
+        )
+        return [cls(index) for index in indices]
 
     @classmethod
     @validate_lonlat
@@ -1243,13 +1391,18 @@ class MOC(AbstractMOC):
 
         Parameters
         ----------
-        list_vertices : list[`~astropy.coordinates.SkyCoord`]
-            A list of `~astropy.coordinates.SkyCoord` objects. Each SkyCoord object should
-            contain more than three vertices and they should each describe a polygon.
+        list_vertices : list[`~astropy.coordinates.SkyCoord`] OR numpy.ndarray
+            - If list_vertices is a list of `~astropy.coordinates.SkyCoord` objects, each
+            SkyCoord object should contain more than three vertices and they should each
+            describe a polygon.
+            - If list_vertices is a numpy.ndarray, it should be in the form
+            [lon_array1, lat_array1, lon_array2, lat_array2, lon_array3, lat_array3, ...].
+            They should be valid longitudes and latitudes in degrees in ICRS.
          max_depth : int, optional
             The resolution of the MOC. Set to 10 by default.
         n_threads : int, optional
-            The number of threads to be used
+            The number of threads to be used. If this is set to None (default value),
+            all available threads will be used.
 
         Returns
         -------
@@ -1264,16 +1417,29 @@ class MOC(AbstractMOC):
         ...     SkyCoord([0, 6, 0, -6], [6, 0, -6, 0], unit="deg")
         ... ]
         >>> list_mocs = MOC.from_polygons(list_vertices)
-        >>> len(list_mocs) # two mocs are created
-        2
+        >>> # without the SkyCoord object, we need to adapt the coordinates
+        >>> list_vertices = [[356, 4, 4, 356], [4, 4, -4, -4],
+        ...                  [0, 6, 0, 354], [6, 0, -6, 0]]
+        >>> list_mocs_no_check_no_wrap = MOC.from_polygons(list_vertices)
+        >>> list_mocs == list_mocs_no_check_no_wrap
+        True
 
         """
-        lon_lat_list = [
-            f(x)
-            for x in list_vertices
-            for f in (lambda x: x.icrs.ra.deg, lambda x: x.icrs.dec.deg)
-        ]
-        indices = mocpy.from_polygons(lon_lat_list, np.uint8(max_depth), n_threads)
+        if isinstance(list_vertices[0], SkyCoord):
+            lon_lat_list = [
+                f(x)
+                for x in list_vertices
+                for f in (lambda x: x.icrs.ra.deg, lambda x: x.icrs.dec.deg)
+            ]
+            indices = mocpy.from_polygons(lon_lat_list, np.uint8(max_depth), n_threads)
+        else:
+            # This is the unsafe version where the users should provide correct coordinates
+            # without checks on our side
+            indices = mocpy.from_polygons(
+                np.array(list_vertices, dtype=np.float64),
+                np.uint8(max_depth),
+                n_threads,
+            )
         return [cls(index) for index in indices]
 
     @classmethod
@@ -1351,7 +1517,8 @@ class MOC(AbstractMOC):
         >>> from mocpy import MOC
         >>> moc1 = MOC.from_stcs("Circle ICRS 147.6 69.9 0.4", max_depth=14)
         >>> moc2 = MOC.from_cone(lon=147.6 * u.deg, lat=69.9 * u.deg, radius=Angle(0.4, u.deg), max_depth=14)
-        >>> assert (moc1 == moc2)
+        >>> moc1 == moc2
+        True
         """
         index = mocpy.from_stcs(stcs, np.uint8(max_depth), np.uint8(delta_depth))
         return cls(index)